Studio Part 1 Works!!!

+++ src/main/java/org/launchcode/controllers/CheeseController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import javax.validation.Valid;
+import javax.validation.constraints.NotNull;
 
 /**
  * Created by LaunchCode
@@ -64,6 +65,9 @@ public class CheeseController {
 
     @RequestMapping(value = "remove", method = RequestMethod.POST)
     public String processRemoveCheeseForm(@RequestParam int[] cheeseIds) {
+
+        if(cheeseIds == null && cheeseIds.length == 0)
+            return "redirect:/cheese";
 
         for (int cheeseId : cheeseIds) {
             cheeseDao.delete(cheeseId);
